refactor webpack plugins and move to prod-only
