+++ package.json
@@ -7,11 +7,11 @@
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
-    "desktop": "webpack --config webpack.dev.js && electron main.js",
+    "desktop": "webpack --config webpack/webpack.dev.js && electron main.js",
     "start": "react-native start",
-    "web": "webpack-dev-server --open --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js",
-    "desktop-build": "webpack --config webpack.prod.js && electron-builder --dir --config electron.config.json",
+    "web": "webpack-dev-server --open --config webpack/webpack.dev.js",
+    "build": "webpack --config webpack/webpack.prod.js",
+    "desktop-build": "webpack --config webpack/webpack.prod.js && electron-builder --dir --config electron.config.json",
     "test": "jest",
     "lint": "eslint .",
     "deploy-ios": "npm version prerelease && git push && fastlane ios beta",
